backend/requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
opencv-python-headless
mediapipe
librosa
python-multipart
joblib
soundfile
torch
kagglehub
python-dotenv
pymongo
openai
